requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cryptg==0.4.0
Flask==2.3.2
frozenlist==1.3.3
gunicorn==20.1.0
hachoir==3.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mutagen==1.46.0
Pillow==9.5.0
pyaes==1.6.1
pyasn1==0.5.0
pycryptodomex==3.18.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
Telethon==1.28.5
urllib3==2.0.3
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
yt-dlp==2023.6.22
